# openapi/media-svc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/media-items": {
      "post": {
        "operationId": "MediaItemController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMediaItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaItem"
                }
              }
            }
          }
        },
        "tags": [
          "media-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "MediaItemController_findAll",
        "parameters": [
          {
            "name": "text",
            "required": false,
            "in": "query",
            "allowEmptyValue": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "explode": true,
            "allowEmptyValue": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "media-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/media-items/{mediaId}": {
      "put": {
        "operationId": "MediaItemController_update",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMediaItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaItem"
                }
              }
            }
          }
        },
        "tags": [
          "media-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "MediaItemController_remove",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "MediaItemController_findOne",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaItemDto"
                }
              }
            }
          }
        },
        "tags": [
          "media-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/media-items/popular": {
      "get": {
        "operationId": "MediaItemController_findPopular",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "media-items"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Media Service",
    "description": "Media Service",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "local"
    },
    {
      "url": "https://media-api.dev.afehrpt.com",
      "description": "staging"
    },
    {
      "url": "https://media-api.afehrpt.com",
      "description": "production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "TagKeyValue": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "value"
        ]
      },
      "CreateMediaItemDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "QZBiXIwwIyWZGhlY1zdKRdW8oaEjYdFMvoamdWfltWcWzD9w02OXRQEtw9CSnMUB0BZ1TGfJa079RFiErgpWDSke5QLnboA1JlF02PzlSKVOL8DfDn79iUSbS4jzX",
            "maxLength": 255,
            "minLength": 0
          },
          "userId": {
            "type": "string",
            "example": "R6XyV2a4jawNlE6UEOqzrup8zU8fWFztgD2KqY9W1cAYo2sqq8KeBWnFCd0OFv9bZbiFDwEwWfqcdc5jHIq7sBAwp9BWV0OgfAncSxi5ULHFBhMiG5whJcDHEVpvJ",
            "maxLength": 255,
            "minLength": 0
          },
          "title": {
            "type": "string",
            "example": "gkNg8mhO7KwpTBRMqsMoEA4fS001eTiViwvnA9Ziw5hLlYbGhJCRwt2uHNTsh1FUg8RrYneDCKZfXfWd0PmetMUPHLgnZFhrL90pg1BMsOQ6BHK9fDvNyWhmaDHiC",
            "maxLength": 255,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "example": "YN7P8AiSA9xB0vzTFa3yX5tRPGNgvSO5wumqeUAvdlqOLfAIKs56n0jxBGyrmCTr1VQ0hmKDzWiAkxjYlL0dbYBAxFqRdTsXXO6LtdXkChaszy41sBXIbSPC8vYF8h1RSQEmURfHE22w9wzdrN3ndAJ8ewZZBpdA4ewJKkimkfK0BOtS1XJHumnAhKY6XlvBF7j5LM1acOpxbgg3Gs5zZeYfJByGBGocf2fpQi4HirZxJ99eLJox6xglH7lh38hRnIBS6HjvgCXR9EGC7ndaM0Xlw266cp7ba1Yl0vkNVNg8"
          },
          "description": {
            "type": "string",
            "example": "AYzzeOSy2CH7bCkM3PlY6O3SUjn2w8nmJCPL9UegP2S2GNjk1um6nMmlubKS5TSCrWlTue6xevTCPggvljdhasm14I8j2TiENeAezrPxg68s6rEbBMtspvkScNcvtlkegrklw2ZUXMHIVlKM3Fl9d0EzUIhWvcrmwoVOEourPjO76ZZJTAqa5aMLQ2xWBRi4gWsV7XAUKRnQBEUucWEfHvOfVTTUx6fNpHD4okSNiNL5oAgfNpKaTGt44EWsUbFSiwIsxODgoUXItSW9701X6dQtyzmvRYMZnkKAWHVnm0h6"
          },
          "uri": {
            "type": "string",
            "example": "9e4nRg0ytp3000mB5ZmrJMXBsLoFgBO6OqGbF5ioAhLWWtGeVXQtt3XsQGhl3tvbuFbbfydGmLnB3Pt55ASxJVQVtcIXIJCRA4toV1p6bziitF4jhrtMMdszM0ccjHVB3np8sk1MSzPTpBzAUifJzYdQWo2jXjblZ1jUaqDEV06W0CKxFF8yOWdahCe4vwLbeHXeyqiwN8jh42zmM4livoeRjR6jWm7elClsWnaqZmAYdH59kuzjCc1fLxxc9n4rUPZIu80Um0muH65RDJYcXYVFaGig9cvgxfUbL3JBemHk"
          },
          "imageSrc": {
            "type": "string",
            "example": "4XaTaxnSqKgLEeCTj4wyWmdlNxb0qd74WSN2B60UHAbwOhsheH8romClmbgNyjjm5wHHjuA3fJ5S9CK1cxjmt0igEUiWS8hMa18heJ9eZRUgMZ640wKejotAQPM7L6LSrnUKKt5v6zsivkBsGG6a4aVRqF2cz5sVVBmM9kWtkJjpePTIjiTz7mtU6EjtC4r3i6E4tkckhPeyzzfyfd87otVYTpU6FK8iaHIkAErPsyWlPm9X1gUlaj9yhVTiOhWchAudLapTGktutwe5RNXDtWoRxOzYwfFS6vpbyoJzcCYZ"
          },
          "isPlayable": {
            "type": "boolean"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "shared",
              "subscription",
              "public"
            ]
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          },
          "eTag": {
            "type": "string",
            "example": "E9iq82hxRuTmisLqmCIS9nqWLB082w7KDL4RKBSNlrvRMOfKkOOG772VD2Y4DiDV1PLlWRuVNklqzExNwWVer64fH75H5QD2pvoZI0Ngw4BkA1LUncp3ofQMtIFEd",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "key",
          "userId",
          "title",
          "description",
          "uri",
          "visibility"
        ]
      },
      "MediaItem": {
        "type": "object",
        "properties": {}
      },
      "MediaVisibilityType": {
        "type": "string",
        "enum": [
          "private",
          "shared",
          "subscription",
          "public"
        ]
      },
      "UpdateMediaItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "y9HcOXm6UtdgOYosqc9zAS2BYw6wE4kELmDy52RKsji6SchkpDboyJYVdBwEfRqMMMyTwrO3x3aiokeh3g8atGf6eboFGqJ2AFGMa04akDxOwqUTyYx21UUmuTw47",
            "maxLength": 255,
            "minLength": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "title": {
            "type": "string",
            "example": "iEiha87f9cMktxeSxqWqXUPzBSc7V2P8AsCPufhi48rJ4I9arQJWLWjxprNehQO0XpzWIptxRoJ9RQXFiK73Q2k6Ld2daxeXFiEvhDORh7LMh3CAs81WjRgz0cs8A",
            "maxLength": 255,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "example": "8lqGDBPtG5kSFMZD77KbYHQbnI9XIV8hO4GclNQQzhzVGXD35vKJ6Nyl5rj5pBPDtaRZNrd7Fk42BiA6haLWVKxHJLK04yGEGOpj94c5HIzdCEGhD88Upl9cNPAWLe9GAARy3vR0Rk5FWRvkeLYJZhQITQZ0U6UY5z9NmLgpFkUPBtSYwMJforIPNCmHOks3zToq9LrMVvrPAAOlrIPD3ke3nCJQHETIMHXez65Xz7cOM1TtGRUvqMqoQhidCk0DElf4reFtHAxf2aW8szKzogr0lVSBUcuntKaupI53Cv9A"
          },
          "description": {
            "type": "string",
            "example": "TzDUbb9tiw9DRCdWZ7tKfEcVASwYk846JLfay6mswQpxcmKTBl215NhWINuWmISyXsCn443ngYQYLGq5um6j30soQm763sV76sRDSf8l8mgqn4LbzlIdfFobzI65g3xvBSMWwGExv1JjNK3hur0lQBPTsxjYgkH88CGhBUwSIxhlXzXcTkkleGzaM3QJ0BgHDMGtdCiivwbZdAd7BOrXuewUnfqM8vPwcb2hwNUrtpDnYgNvrbtMwnzPlLAHalx5lMjyL3r7vArOiZ4hjVkYj1UQcdsNIAx99jGbb5sWLL1Y"
          },
          "imageSrc": {
            "type": "string",
            "example": "MzNaGrQAnhuWRTTieKvqttm3SUfh6pTiJUfJWbeFtkDBKbcrEt8SBuicsp5VRX8by4FApsVAFpVsfiO9XqZ0ooWDEHxOnFo1TkyeUtWl3CPJ9kZdWutQTjw2Szw6ekgsNVZQiEThAvtB6kRiVcAmBZBoBqSkM5vo9KpbtscjtbiqJkt2Jukhnk1IumVGINsPUMiPlVjBNyeEF6fGwdfOg2Qf8egJ5yDJU02HF2jxPPnfdYqvTOECpQIIDv7AY2bgCZzXEujbvcTBfk21X3xmncPY0InBDZu2kraBme1DD9a7"
          },
          "visibility": {
            "$ref": "#/components/schemas/MediaVisibilityType"
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          },
          "eTag": {
            "type": "string",
            "example": "XK7Xp1c55oQiJ6jTzIUp5hHXr4cb41yx7BiBrKkhrcfjgFWjLxeuGSEmq1gxRfp69tZ4O8KCAn3sprRbwD3KEBH9kTOclX8pLHFzdQSSOv8I0mEd4PBhC9Bsh673j",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "_id",
          "title",
          "description",
          "visibility"
        ]
      },
      "MediaItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "y9HcOXm6UtdgOYosqc9zAS2BYw6wE4kELmDy52RKsji6SchkpDboyJYVdBwEfRqMMMyTwrO3x3aiokeh3g8atGf6eboFGqJ2AFGMa04akDxOwqUTyYx21UUmuTw47",
            "maxLength": 255,
            "minLength": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "key": {
            "type": "string",
            "example": "hk5NOHrpKvyqlnACOMld8DQb3y2Ta4Zu8fS3IzNiGVPeuhqEv1NwD8dTLQXkUrbD6nVHJLew5Y0u3Q5LFWdeEuTOQUhHP06jivSD09iQaQi8VCJQVIpbxApoDQnyA",
            "maxLength": 255,
            "minLength": 0
          },
          "userId": {
            "type": "string",
            "example": "7mvTNZX6wYVJuQUrunHlpoUjqrUXziIPq3n5xgyCFdm53NJO2DluqgEILWyzwGuY91U4EVTC5qzq9mFjhhaLuOwc6k9KXwBR1gfuvS1du6H3vB8Cn8ROFhPHni1YS",
            "maxLength": 255,
            "minLength": 0
          },
          "title": {
            "type": "string",
            "example": "hBKzoS8DyCHy6ybVnUOihbCGNArHs6utlRkKIwnujs05LDlG75cMjn8IGpIMFDtdjIBnF7SVSJXyKUFJoXwwJp1U5lIE8g2aCpKKew5bN2IHxSCLRuOHlzY8xwS45",
            "maxLength": 255,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "example": "KKaZ0UaSY3K8YhuzwcahA70SX9TUchbiy38JZPQWFmZf3azg23c7A6JY6qIMqhpKxr5PGyAuP1kfgPifYrk7vBccyAEQexXXDN0KPW54EN2rAwtTuExfF7yD9Ubh1JiSQmJksYpbYdbTJDZJJhowLq8UnJfpEa7o9qWZlV7avZlcjWc7x62UIbGHx2bVTe3NhSLJExdc1Eik5BLaKwTFy7WFbSy8NcGiZA3I0hUflhoggj8FUCJ59XWW51fRiB3TrDrGdPhrY3BF8kWlaS1Ow3pOgffXdp0FVtRLpX1khkxx"
          },
          "description": {
            "type": "string",
            "example": "1AMrpE8mixH5tnZ7wLRAULuFzeUqNu4EeZohF1v1a7Z8nsxuADX0ilcfXV3dGjDcJXrS2ZavUKliDte9tZHF8b1tT0EdjNARdJxIAT5Z1slPVI0uexAwY1jOvXB0RT0DuULF4Bquma5TH9FAROI0Afi1iN5J4CuknXaMpU2kmLMHr1kDsmlPflOW4DjPbjRBL2ecG4DtDiArZ3zPaE16R1vjCkDaEg1WZ1hQufTX1xjSKRolDRK0GoxqrckyOJ79nCflkiDyzp8aFqZdTmjK3lZBfjIC19IkfhV2gN9RDHeV"
          },
          "uri": {
            "type": "string",
            "example": "jIQSOVTSCbQGZEIVkCWv5W5zhxsfzY8jBj98C3zb2GsAboCJ3MdmMh0P6A9L2zUyDqsmsozU9A84s3Sm61SGBCpHgOA78IiLU3HVmQ4TeIcEG0gGGtrkrW3wx7rwZuZclVW50QmVRUr9hmTmNiqTifJmsf5dDlRDIVgxxzwF4VaG53COK4zItVbvdzljUWoPfIaBVxRpVTdTkMzw2B3GjHupcUkdbQMqraKKYhjaUXkMSIdWeY4lzh131sZquVg3uSoE2qteYOIbP49oTuESp3ZoZpNwYzsZMjL178UbygdP"
          },
          "imageSrc": {
            "type": "string",
            "example": "7TJuQYS4yiKTFVNBygRTr72pCeIkrFFtEFSf8aBYE6ymk4ivX6yMescCepSgKJmifiZEBZiwMgsVRFfZ5Anm3EkTly2TpCU65FeFJNlRNkufi4lRPbAHzqxWAfm0f6nlFdH9ha1i8n9s0S54q6P0ZIoilFn4DjoEfqm6ODwPuuV77zqQkE648v4mjsq1poPWx7V3bbAo2pl6LWZICEP1nc9q5zBK440vV3liahgrR26lbw4QXAosizs4jUZechUo7hjju4zy5SLfJc28jzUFWl0zpwEnxC5v64vzIjMUGnRC"
          },
          "isPlayable": {
            "type": "boolean"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "shared",
              "subscription",
              "public"
            ]
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          },
          "eTag": {
            "type": "string",
            "example": "Zu2uSsVttocQw7iuPU418qu1Y5P4rI8Ch9QIyYtb0OO4qAefJgtKQowYLNpTpktezO4pkdkDcjPB6EfeLHhs4dbTZaRhpfJDzWOBEiVHt8gyOHwD2qDSBlvA29egG",
            "maxLength": 255,
            "minLength": 0
          },
          "shareCount": {
            "type": "number"
          },
          "viewCount": {
            "type": "number"
          },
          "likesCount": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "key",
          "userId",
          "title",
          "description",
          "uri",
          "visibility",
          "shareCount",
          "viewCount",
          "likesCount"
        ]
      },
...
    }
  }
}